[INT-204] Validate CRUD methods (#74)
* [INT-204] Validate CRUD methods

Add validation to CRUD methods for CarePlan, CareTeam, and Task, only allowing valid participants of a CarePlan to access resources related to that CarePlan

* Validate Task fields on Create

* Validate Task fields on Update

* Validate Get for CareTeam/CarePlan/Task

* Add support for parameters when getting CarePlan

* Update orchestrator/careplanservice/handle_get.go

Co-authored-by: reinkrul <info@reinkrul.nl>

* Add basic CRUD auth flow diagram

* Address PR refactors of Update and Get

* Fix merge conflicts

* Added unit test for creating Task when CarePlan doesn't exist

* Refactor Get and Search, got started with unit test implementation

* Address PR comments, fix e2e unit test, implement Search for Task and CareTeam

* Implement basic unit tests for CRUD methods, address PR comments

I discovered during implementation of unit tests that it is very tricky to mock the return of a side effect even with reflection, we should refactor the unit tests to http tests but that is a lot of work so it can be out of scope for now as the new functionality is covered by integ tests. A new task will be created to track this

* Implement URA check when creating a new CarePlan and CareTeam

This breaks the e2e test, this is due to a Nuts bug. Will keep working on this

* Update Task validation logic to include owner and requester, fix e2e test, only local care provider can create new CarePlans and CareTeams

* Fix failing unit test

---------

Co-authored-by: reinkrul <info@reinkrul.nl>