DEV-19 | Implement azure adb2c token client
Adds a client that can parse an ADB2C JWT, validate it is not expired, validate it contains required claims, and compares it against a list of known issuers to ensure the token comes from a source we trust.

Automatically handles key rotation for discovery endpoints.

This will be added to CPC in a future commit, for now this is a self-contained change, with extensive unit tests and an integration test that can be used with a real JWT and discovery endpoint